common: add crypto random alternative
* ctr_drbg was using mbedtls_entropy_func as entropy
source. Replace it with strong_entropy_func, which uses
hardware TRNG to gather a seed.
* An entropy context struct is no longer needed.

Notes:

Couldn't overwrite only otPlatCryptoRandomInit,
since it has to reference the same drbg context as
otPlatCryptoRandomDeinit and otPlatCryptoRandomGet.

mbedtls_entropy_func is using an accumulator for entropy.
It starts a SHA256 process (1) to update the current accumulator,
but it also starts another SHA256 process (2), before the previous
one is finished, to compute the output hash, which will be the seed.
Furthermore, (1) finishes at the next call of mbedtls_entropy_func.
When using hardware SHA256, this could block access to it, since we may
run into a case when mbedtls_entropy_func is not called anymore, thus
the SHA256 mutex will never be unlocked. This can happen in both Matter
and OT, if usage of hardware SHA256 is enabled.

Behind the scenes, this function was also called through handleMbedtlsEntropyPoll,
from crypto_platform.cpp. So each call to mbedtls_entropy_func would generate at
least a call to otPlatEntropyGet. In conclusion, the call frequency should be similar
to mbedtls_entropy_func usage, so no additional risk of entropy depletion is introduced.

Additional relevant work can be found here:
https://bitbucket.sw.nxp.com/projects/CONNINT/repos/connectivity-framework/pull-requests/2912/overview

Signed-off-by: Marius Tache <marius.tache@nxp.com>