k32w0: RCP: config: NCP_HDLC_TX_CHUNK_SIZE
Always make sure that the hdlc encoder is bigger that the NCP spinel TX buffer.
In fact some issues could happen if the HDLC encoder buffer is smaller that the spinel TX buffer.
So to avoid such issue we always make sure that the hdlc encoder is bigger that the NCP spinel TX buffer.