k32w0: uart: Aligning OT_PLAT_UART_RX_BUFFER_SIZE with the size of the serial manager rx ring buffer
Fix issue when the serial rx ring buffer is bigger than the ot plat UART buffer.
In fact there is a risk to not re-schedule the serial manager task if there is more byte in the rx ring buffer than what ot uart could read. To avoid such an issue always align the two buffer size so that the whole ring buffer could be read in one run.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>