K32w0 sept 2022 (#352)
* common: add crypto random alternative

* ctr_drbg was using mbedtls_entropy_func as entropy
source. Replace it with strong_entropy_func, which uses
hardware TRNG to gather a seed.
* An entropy context struct is no longer needed.

Notes:

Couldn't overwrite only otPlatCryptoRandomInit,
since it has to reference the same drbg context as
otPlatCryptoRandomDeinit and otPlatCryptoRandomGet.

mbedtls_entropy_func is using an accumulator for entropy.
It starts a SHA256 process (1) to update the current accumulator,
but it also starts another SHA256 process (2), before the previous
one is finished, to compute the output hash, which will be the seed.
Furthermore, (1) finishes at the next call of mbedtls_entropy_func.
When using hardware SHA256, this could block access to it, since we may
run into a case when mbedtls_entropy_func is not called anymore, thus
the SHA256 mutex will never be unlocked. This can happen in both Matter
and OT, if usage of hardware SHA256 is enabled.

Behind the scenes, this function was also called through handleMbedtlsEntropyPoll,
from crypto_platform.cpp. So each call to mbedtls_entropy_func would generate at
least a call to otPlatEntropyGet. In conclusion, the call frequency should be similar
to mbedtls_entropy_func usage, so no additional risk of entropy depletion is introduced.

Additional relevant work can be found here:
https://bitbucket.sw.nxp.com/projects/CONNINT/repos/connectivity-framework/pull-requests/2912/overview

Signed-off-by: Marius Tache <marius.tache@nxp.com>

* common: ram_storage: fix expression and ramDescrHeaderSize
 * fix if directive expression
 * compute kRamDescHeaderSize based on offset

Signed-off-by: Andrei Menzopol <andrei.menzopol@nxp.com>

* k32w0: pdm_ram_storage_glue: fix resize, add macro
 * fix buffer overflow and comparison in ram buffer resize function
 * change offsetof to kRamDescHeaderSize in get ram buffer function

Signed-off-by: Andrei Menzopol <andrei.menzopol@nxp.com>

* common: fixing freertos alarm timer implementation

Fixing wrong usage of xTimerGetExpiryTime

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>

* k32w0: RCP: config: NCP_HDLC_TX_CHUNK_SIZE

Always make sure that the hdlc encoder is bigger that the NCP spinel TX buffer.
In fact some issues could happen if the HDLC encoder buffer is smaller that the spinel TX buffer.
So to avoid such issue we always make sure that the hdlc encoder is bigger that the NCP spinel TX buffer.

* k32w0: build: switching the K32W0 RCP UART baudrate to 1M

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>

* k32w0: uart: Aligning OT_PLAT_UART_RX_BUFFER_SIZE with the size of the serial manager rx ring buffer

Fix issue when the serial rx ring buffer is bigger than the ot plat UART buffer.
In fact there is a risk to not re-schedule the serial manager task if there is more byte in the rx ring buffer than what ot uart could read. To avoid such an issue always align the two buffer size so that the whole ring buffer could be read in one run.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>

* k32w0: radio: state machine update

- check for ongoing Rx before Tx

- before radio state change, stop the radio
  with vMMAC_RadioToOffAndWait() so there are no pending interrupts

- don't restart Rx on the same channel
- check for ongoing Rx before starting Rx on a different channel

- allow otPlatRadioSetTransmitPower() only in sleep / Rx states

- handle Rx events independently in K32WISR()

- add OT_RADIO_CAPS_SLEEP_TO_TX

- configurable rx ring (K32W0_RADIO_NUM_OF_RX_BUFS) - must be power of 2

- after tx ends, rx on the same channel: improves poll scenario in RCP context after energy scan

- don't Rx when no free frame
    - added OT_RADIO_STATE_RX_DISABLED

    - K32WEnableReceive(): transition from rx state to rx_disabled when there are no frames available.
      K32WEnableReceive() is called after rx ring state changes or when radio is idle.

    - reworked the Rx ring buffer: lockless operations
      full state: head + 1 == tail
      empty state: head == tail

    - rx overflow counter

- rename otPlatRadioGetCslClockUncertainty() to otPlatRadioGetCslUncertainty()
- set CSL_UNCERT to 255 to maintain the same CSL settings that were used for the certification

- clear sRadioInitForLp after wakeup
- don't config sTxOtFrame since it's set by the stack before tx

Signed-off-by: George Stefan <george.stefan@nxp.com>

Signed-off-by: Marius Tache <marius.tache@nxp.com>
Signed-off-by: Andrei Menzopol <andrei.menzopol@nxp.com>
Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>
Signed-off-by: George Stefan <george.stefan@nxp.com>
Co-authored-by: Marius Tache <marius.tache@nxp.com>
Co-authored-by: Andrei Menzopol <andrei.menzopol@nxp.com>
Co-authored-by: Gatien Chapon <gatien.chapon@nxp.com>