k32w1: uart: allow DMA usage for UART
In some cases (i.e. RCP) where the connections do not allow
for reliable FC propagation from host to MCU, the LPUART DMA
mechanism can be used. This ensures the reliability of the
UART communications over spinel, which is mandatory for
scenarios like border routers.

As such this patch adds the glue defines and structures
so that the user can select from the target to be built that
DMA is desired, as opposed to the standard polling mechanism.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>