k32w1: build: add RCP DMA only target
This patch adds a specific target that allows the user
to remove the FC pins required for reliable UART communications
and rely on the DMA mechanism present within the platform.
This target is to be used when the device is used as a dongle
providing OT services to a Linux host.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>