[lib] fix lib config build issue (#792)
This commit fixes the build issue of rt1060 when bumping openthread to
`d0fbfb8`.

The root cause is that the config file of `openthread/lib` has been
independent of the OT core config file. From now if we want to
configure for variables used in `lib` (OPENTHREAD_LIB_XXX) we need to
define it in another lib config file. (Can be specified in cmake by
`OT_LIB_CONFIG`).

The `Init` method of `RadioSpinel` is also updated. The PR updates the
usage in platform radio code.